UPD: Don't deactivate panel when losing focus, but only when another panel is activated. This reduces flickering when switching tabs if brightness level of inactive panel setting is used.
The deactivation on losing focus was done because some actions were being disabled when panels did not have focus (like Ctrl+X, Ctrl+C). But now that this method has been abandoned in favour of possibility of assiging hotkeys to specific controls, the deactivation is not needed. This reverts to having always one panel highlighted.